Function signature changed consistent with mysql

+++ integ-test/src/test/java/com/amazon/opendistroforelasticsearch/sql/ppl/MathematicalFunctionIT.java
@@ -51,7 +51,7 @@ public class MathematicalFunctionIT extends PPLIntegTestCase {
         executeQuery(
             String.format(
                 "source=%s | eval f = ceil(age) | fields f", TEST_INDEX_BANK));
-    verifySchema(result, schema("f", null, "long"));
+    verifySchema(result, schema("f", null, "integer"));
     verifyDataRows(
         result,
         rows(32), rows(36), rows(28), rows(33), rows(36), rows(39), rows(34));
@@ -75,7 +75,7 @@ public class MathematicalFunctionIT extends PPLIntegTestCase {
         executeQuery(
             String.format(
                 "source=%s | eval f = floor(age) | fields f", TEST_INDEX_BANK));
-    verifySchema(result, schema("f", null, "long"));
+    verifySchema(result, schema("f", null, "integer"));
     verifyDataRows(
         result,
         rows(32), rows(36), rows(28), rows(33), rows(36), rows(39), rows(34));
